backend/requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
httpx

# PDF Processing
PyMuPDF

# AI and Embeddings
openai
pydantic-ai

# Vector Database
pinecone

# Streamlit Frontend
streamlit

# Utilities
numpy
python-multipart
langchain
tiktoken 
